Level check for groupmates disabled
