+++ src/skills/relocate.h
@@ -1,0 +1,10 @@
+#ifndef BYLINS_RELOCATE_H
+#define BYLINS_RELOCATE_H
+
+class CHAR_DATA;
+
+void do_relocate(CHAR_DATA *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_RELOCATE_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
